requirements.txt
pandas
psycopg2-binary
python-dotenv
streamlit
numpy
altair<5
